+++ src/buildtool/build_engine/target_map/built_in_rules.cpp
@@ -1369,8 +1369,17 @@ void GenericRuleWithDeps(
     for (auto const& dep : dependency_values) {
         inputs = ExpressionPtr{Expression::map_t{inputs, (*dep)->RunFiles()}};
     }
+    auto inputs_conflict = BuildMaps::Target::Utils::tree_conflict(inputs);
     for (auto const& dep : dependency_values) {
         inputs = ExpressionPtr{Expression::map_t{inputs, (*dep)->Artifacts()}};
+    }
+    // While syntactical conflicts are resolved in a latest wins (with artifacts
+    // after runfiles), semantic path conclicts are an error.
+    if (inputs_conflict) {
+        (*logger)(fmt::format("Input artifacts have staging conflict on {}",
+                              nlohmann::json(*inputs_conflict).dump()),
+                  /*fatal=*/true);
+        return;
     }
     std::vector<Tree::Ptr> trees{};
     inputs = BuildMaps::Target::Utils::add_dir_for(